# chat-application/src/app/data/api/models/register-user-post-wrapper.ts
/* tslint:disable */
export interface RegisterUserPostWrapper {

  /**
   * Email
   */
  email: string;

  /**
   * Password
   */
  password: string;

  /**
   * Username
   */
  username: string;
}